refactor(labelmap): component renaming
