+++ src/components/SegmentList.vue
